racing-game/requirements.txt
numpy
tensorflow-cpu ~= 2.4.0
# ---- Added by the Emily CLI ----
enum34
